+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Bug fixes
 
+### Enhancements
+- TT-1798 Implement staff sso
+
 ## [2019.10.22](https://github.com/uktrade/directory-sso-proxy/releases/tag/2019.10.22)
 [Full Changelog](https://github.com/uktrade/directory-sso-proxy/compare/2019.07.09...2019.10.22)